Move custom exceptions to exceptions, fix circular import
